remote config nested class 이용
