--- app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/config/SettingsBadgeConfig.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.snutt2.lib.network.dto.config
-
-import com.squareup.moshi.Json
-
-data class SettingsBadgeConfig(
-    @Json(name = "new") val new: List<String> = emptyList()
-)
