--- app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/config/VacancyBannerConfig.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.snutt2.lib.network.dto.config
-
-import com.squareup.moshi.Json
-
-data class VacancyBannerConfig(
-    @Json(name = "visible") val visible: Boolean = false,
-)
