--- app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/config/VacancyUrlConfig.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.snutt2.lib.network.dto.config
-
-import com.squareup.moshi.Json
-
-data class VacancyUrlConfig(
-    @Json(name = "url") val url: String? = null,
-)
